only clean icons for thing that aren't full commands

+++ challenge/docker-entrypoint.d/33_remove_tool_shortcuts.sh
@@ -4,7 +4,7 @@ for LAUNCHER in /home/hacker/.config/xfce4/panel/launcher-*
 do
         NUMBER=${LAUNCHER##*-}
         EXE=$(cat "$LAUNCHER"/*.desktop | grep Exec | sed -e "s/.*=//")
-        if [ ! -f "$EXE" ]
+        if [ ! -f "$EXE" ] && [[ ! "$EXE" =~ " " ]]
         then
                 rm -rf "$LAUNCHER"
                 sed -i -e '/<property name="plugin-$NUMBER" type="string" value="launcher">/,/<\/property>/d' /home/hacker/.config/xfce4/xfconf/xfce-perchannel-xml/xfce4-panel.xml
